meetwise-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
aiofiles==23.2.1
motor==3.2.0
pydantic==2.5.0
PyJWT==2.8.0
pymongo==4.6.1
email-validator==2.1.0.post1 